Move common props functions to util
