+++ lib/rules/boolean-prop-naming.js
@@ -6,6 +6,7 @@
 
 const has = require('has');
 const Components = require('../util/Components');
+const propsUtil = require('../util/props');
 
 // ------------------------------------------------------------------------------
 // Rule Definition
@@ -48,26 +49,6 @@ module.exports = {
 
     // Remembers all Flowtype object definitions
     const objectTypeAnnotations = new Map();
-
-    /**
-     * Checks if node is `propTypes` declaration
-     * @param {ASTNode} node The AST node being checked.
-     * @returns {Boolean} True if node is `propTypes` declaration, false if not.
-     */
-    function isPropTypesDeclaration(node) {
-      if (node && node.type === 'ClassProperty') {
-        // Flow support
-        if (node.typeAnnotation && node.key.name === 'props') {
-          return true;
-        }
-        node = node.key;
-      }
-
-      return Boolean(
-        node &&
-        node.name === 'propTypes'
-      );
-    }
 
     /**
      * Returns the prop key to ensure we handle the following cases:
@@ -163,7 +144,7 @@ module.exports = {
 
     return {
       ClassProperty: function(node) {
-        if (!rule || !isPropTypesDeclaration(node)) {
+        if (!rule || !propsUtil.isPropTypesDeclaration(node)) {
           return;
         }
         if (node.value && node.value.properties) {
@@ -175,7 +156,7 @@ module.exports = {
       },
 
       MemberExpression: function(node) {
-        if (!rule || !isPropTypesDeclaration(node.property)) {
+        if (!rule || !propsUtil.isPropTypesDeclaration(node)) {
           return;
         }
         const component = utils.getRelatedComponent(node);
@@ -192,7 +173,7 @@ module.exports = {
 
         // Search for the proptypes declaration
         node.properties.forEach(property => {
-          if (!isPropTypesDeclaration(property.key)) {
+          if (!propsUtil.isPropTypesDeclaration(property)) {
             return;
           }
           validatePropNaming(node, property.value.properties);
