+++ lib/rules/forbid-prop-types.js
@@ -4,6 +4,7 @@
 'use strict';
 
 const variableUtil = require('../util/variable');
+const propsUtil = require('../util/props');
 
 // ------------------------------------------------------------------------------
 // Constants
@@ -45,22 +46,6 @@ module.exports = {
 
       const forbid = configuration.forbid || DEFAULTS;
       return forbid.indexOf(type) >= 0;
-    }
-
-    /**
-     * Checks if node is `propTypes` declaration
-     * @param {ASTNode} node The AST node being checked.
-     * @returns {Boolean} True if node is `propTypes` declaration, false if not.
-     */
-    function isPropTypesDeclaration(node) {
-      if (node.type === 'ClassProperty') {
-        node = node.key;
-      }
-
-      return Boolean(
-        node &&
-        node.name === 'propTypes'
-      );
     }
 
     /**
@@ -147,14 +132,14 @@ module.exports = {
 
     return {
       ClassProperty: function(node) {
-        if (!isPropTypesDeclaration(node)) {
+        if (!propsUtil.isPropTypesDeclaration(node)) {
           return;
         }
         checkNode(node.value);
       },
 
       MemberExpression: function(node) {
-        if (!isPropTypesDeclaration(node.property)) {
+        if (!propsUtil.isPropTypesDeclaration(node)) {
           return;
         }
 
@@ -167,7 +152,7 @@ module.exports = {
             return;
           }
 
-          if (!isPropTypesDeclaration(property.key)) {
+          if (!propsUtil.isPropTypesDeclaration(property)) {
             return;
           }
           if (property.value.type === 'ObjectExpression') {
