+++ lib/rules/prop-types.js
@@ -12,6 +12,7 @@ const Components = require('../util/Components');
 const variable = require('../util/variable');
 const annotations = require('../util/annotations');
 const versionUtil = require('../util/version');
+const propsUtil = require('../util/props');
 
 // ------------------------------------------------------------------------------
 // Constants
@@ -180,22 +181,6 @@ module.exports = {
         }
       }
       return false;
-    }
-
-    /**
-     * Checks if we are declaring a prop
-     * @param {ASTNode} node The AST node being checked.
-     * @returns {Boolean} True if we are declaring a prop, false if not.
-     */
-    function isPropTypesDeclaration(node) {
-      if (node && node.type === 'ClassProperty') {
-        node = node.key;
-      }
-
-      return Boolean(
-        node &&
-        node.name === 'propTypes'
-      );
     }
 
     /**
@@ -969,7 +954,7 @@ module.exports = {
       ClassProperty: function(node) {
         if (isAnnotatedClassPropsDeclaration(node)) {
           markPropTypesAsDeclared(node, resolveTypeAnnotation(node));
-        } else if (isPropTypesDeclaration(node)) {
+        } else if (propsUtil.isPropTypesDeclaration(node)) {
           markPropTypesAsDeclared(node, node.value);
         }
       },
@@ -1006,7 +991,7 @@ module.exports = {
         let type;
         if (isPropTypesUsage(node)) {
           type = 'usage';
-        } else if (isPropTypesDeclaration(node.property)) {
+        } else if (propsUtil.isPropTypesDeclaration(node)) {
           type = 'declaration';
         }
 
@@ -1027,7 +1012,7 @@ module.exports = {
       },
 
       MethodDefinition: function(node) {
-        if (!node.static || node.kind !== 'get' || !isPropTypesDeclaration(node.key)) {
+        if (!node.static || node.kind !== 'get' || !propsUtil.isPropTypesDeclaration(node)) {
           return;
         }
 
@@ -1046,7 +1031,7 @@ module.exports = {
       ObjectExpression: function(node) {
         // Search for the proptypes declaration
         node.properties.forEach(property => {
-          if (!isPropTypesDeclaration(property.key)) {
+          if (!propsUtil.isPropTypesDeclaration(property)) {
             return;
           }
           markPropTypesAsDeclared(node, property.value);
