+++ lib/rules/require-default-props.js
@@ -9,6 +9,7 @@ const Components = require('../util/Components');
 const variableUtil = require('../util/variable');
 const annotations = require('../util/annotations');
 const astUtil = require('../util/ast');
+const propsUtil = require('../util/props');
 
 const QUOTES_REGEX = /^["']|["']$/g;
 
@@ -29,33 +30,6 @@ module.exports = {
   create: Components.detect((context, components, utils) => {
     const sourceCode = context.getSourceCode();
     const propWrapperFunctions = new Set(context.settings.propWrapperFunctions || []);
-
-    /**
-     * Checks if the Identifier node passed in looks like a propTypes declaration.
-     * @param   {ASTNode}  node The node to check. Must be an Identifier node.
-     * @returns {Boolean}       `true` if the node is a propTypes declaration, `false` if not
-     */
-    function isPropTypesDeclaration(node) {
-      return astUtil.getPropertyName(node) === 'propTypes';
-    }
-
-    /**
-     * Checks if the Identifier node passed in looks like a defaultProps declaration.
-     * @param   {ASTNode}  node The node to check. Must be an Identifier node.
-     * @returns {Boolean}       `true` if the node is a defaultProps declaration, `false` if not
-     */
-    function isDefaultPropsDeclaration(node) {
-      return (astUtil.getPropertyName(node) === 'defaultProps' || astUtil.getPropertyName(node) === 'getDefaultProps');
-    }
-
-    /**
-     * Checks if the PropTypes MemberExpression node passed in declares a required propType.
-     * @param   {ASTNode} propTypeExpression node to check. Must be a `PropTypes` MemberExpression.
-     * @returns {Boolean}                    `true` if this PropType is required, `false` if not.
-     */
-    function isRequiredPropType(propTypeExpression) {
-      return propTypeExpression.type === 'MemberExpression' && propTypeExpression.property.name === 'isRequired';
-    }
 
     /**
      * Find a variable by name in the current scope.
@@ -131,7 +105,7 @@ module.exports = {
 
       return props.map(property => ({
         name: sourceCode.getText(property.key).replace(QUOTES_REGEX, ''),
-        isRequired: isRequiredPropType(property.value),
+        isRequired: propsUtil.isRequiredPropType(property.value),
         node: property
       }));
     }
@@ -332,8 +306,8 @@ module.exports = {
 
     return {
       MemberExpression: function(node) {
-        const isPropType = isPropTypesDeclaration(node);
-        const isDefaultProp = isDefaultPropsDeclaration(node);
+        const isPropType = propsUtil.isPropTypesDeclaration(node);
+        const isDefaultProp = propsUtil.isDefaultPropsDeclaration(node);
 
         if (!isPropType && !isDefaultProp) {
           return;
@@ -382,7 +356,7 @@ module.exports = {
           if (isPropType) {
             addPropTypesToComponent(component, [{
               name: node.parent.property.name,
-              isRequired: isRequiredPropType(node.parent.parent.right),
+              isRequired: propsUtil.isRequiredPropType(node.parent.parent.right),
               node: node.parent.parent
             }]);
           } else {
@@ -414,8 +388,8 @@ module.exports = {
           return;
         }
 
-        const isPropType = isPropTypesDeclaration(node);
-        const isDefaultProp = isDefaultPropsDeclaration(node);
+        const isPropType = propsUtil.isPropTypesDeclaration(node);
+        const isDefaultProp = propsUtil.isDefaultPropsDeclaration(node);
 
         if (!isPropType && !isDefaultProp) {
           return;
@@ -522,8 +496,8 @@ module.exports = {
             return;
           }
 
-          const isPropType = isPropTypesDeclaration(property);
-          const isDefaultProp = isDefaultPropsDeclaration(property);
+          const isPropType = propsUtil.isPropTypesDeclaration(property);
+          const isDefaultProp = propsUtil.isDefaultPropsDeclaration(property);
 
           if (!isPropType && !isDefaultProp) {
             return;
