+++ lib/rules/sort-prop-types.js
@@ -4,6 +4,7 @@
 'use strict';
 
 const variableUtil = require('../util/variable');
+const propsUtil = require('../util/props');
 
 // ------------------------------------------------------------------------------
 // Rule Definition
@@ -45,22 +46,6 @@ module.exports = {
     const ignoreCase = configuration.ignoreCase || false;
     const sortShapeProp = configuration.sortShapeProp || false;
     const propWrapperFunctions = new Set(context.settings.propWrapperFunctions || []);
-
-    /**
-     * Checks if node is `propTypes` declaration
-     * @param {ASTNode} node The AST node being checked.
-     * @returns {Boolean} True if node is `propTypes` declaration, false if not.
-     */
-    function isPropTypesDeclaration(node) {
-      if (node.type === 'ClassProperty') {
-        node = node.key;
-      }
-
-      return Boolean(
-        node &&
-        node.name === 'propTypes'
-      );
-    }
 
     function getKey(node) {
       return sourceCode.getText(node.key || node.argument);
@@ -199,14 +184,14 @@ module.exports = {
       },
 
       ClassProperty: function(node) {
-        if (!isPropTypesDeclaration(node)) {
+        if (!propsUtil.isPropTypesDeclaration(node)) {
           return;
         }
         checkNode(node.value);
       },
 
       MemberExpression: function(node) {
-        if (!isPropTypesDeclaration(node.property)) {
+        if (!propsUtil.isPropTypesDeclaration(node)) {
           return;
         }
 
@@ -219,7 +204,7 @@ module.exports = {
             return;
           }
 
-          if (!isPropTypesDeclaration(property.key)) {
+          if (!propsUtil.isPropTypesDeclaration(property)) {
             return;
           }
           if (property.value.type === 'ObjectExpression') {
